Allow alpha for foreground images
Fixed an oversight where I set the foreground image to have alpha set to false. It was set this way after discovering that performance could be significantly increased during recording if the canvas has alpha false. The foreground image should not need this though. Only the canvas that is being recorded.